Make many structures const.
These structures should never be changed on runtime so they should be
marked as constant. This means the compiler can now warn us if we make
the mistake of trying to change any of them, but more importantly, it
gives the compiler more information about the nature of these and
therefor lets the linker map these structures to read-only memory
instead of read-write, which should improve page deduplication in many
cases and reduce overall system memory usage.

This has reduced the number of dirty memory pages from 10 to 2, which
translates to 32KiB of memory saved per process linking to libunibreak
starting from the second process.