Fix marking of mid-char (utf8/16) with regional indicators.
There was a bug causing those not to be marked correctly when using
regional indicators. An example broken text would be: "═🇫".

This was due to a forgotten call to the function in charge of marking
the breaks, causing it to default to 0 (break) instead of having 2
(inside char) set.

Fixes #15.